support formatting in the playground
