+++ package.json
@@ -13,6 +13,7 @@
   },
   "packageManager": "pnpm@9.0.4",
   "dependencies": {
+    "@dprint/formatter": "^0.3.0",
     "@effect-rx/rx-react": "^0.28.16",
     "@effect/platform": "^0.53.14",
     "@effect/platform-browser": "^0.33.29",
